+++ src/src/utils.h
@@ -12,7 +12,7 @@ long rng(void);
 
 void rngSeed(long newSeed);
 // 0..255 returned
-long rng8Bit(void);
+uint8_t rng8Bit(void);
 // 0..31 returned
 long rng5Bit(void);
